Clean up OrderedHashTable Resizing https://bugs.webkit.org/show_bug.cgi?id=297708 rdar://158014775
Reviewed by Yusuke Suzuki.

When inserting into an OrderedHashTable, if the buffer is resized then subsequently an
exception is thrown. The original buffer will be left in a inconsistent state. This patch
takes two approaches to this problem:

1) Compute the hash key before resizing. This isn't observable semantically since resizing
can only OOM and that is free to happen anywhere.

2) Only commit the new buffer once everything is set up. This way if we change our design
in the future and throw somewhere in the middle we'll just roll back to the previous state.

Lastly, convert some shifts into multiplications/divisions by the appropriate power of two.
Clang will convert the muls/divs into the corresponding shifts so we should focus on
readability.

Canonical link: https://commits.webkit.org/297297.293@safari-7622-branch