[Cocoa] Fix thread safety issues with the Web Share API https://bugs.webkit.org/show_bug.cgi?id=298063 rdar://159066364
Reviewed by Abrar Rahman Protyasha and Ryosuke Niwa.

Ensure use of `UIActivityItemProvider` and `NSPreviewRepresentingActivityItem`
happens only on the main thread. These APIs are not thread-safe.

* Source/WebKit/UIProcess/Cocoa/WKShareSheet.mm:
(appendFilesAsShareableURLs):

Make a slight behavior change to avoid `nil`-ing out the entire share
data array if an item for a file could not be successfully created. In this
scenario, the title/text/URL can still be shared.

Canonical link: https://commits.webkit.org/297297.358@safari-7622-branch