ASAN_ILL | WebCore::KeyframeEffectStack::allowsAcceleration; WebCore::KeyframeEffectStack::removeEffect; WebCore::KeyframeEffect::updateEffectStackMembership https://bugs.webkit.org/show_bug.cgi?id=298317 rdar://157023038
Reviewed by Cameron McCormack.

When a document is torn down, we call into `AnimationTimelinesController::detachFromDocument()` which
will call `AnimationTimeline::detachFromDocument()` for each known timeline and then `WebAnimation::remove()`
for each animation associated with that timeline. This will eventually disassociate animation effects
from the `KeyframeEffectStack` of their target, ensuring there are no null references in the stack's
list of effects.

However, with the advent of scroll-driven animations, an animation's timeline may be set to null, which
will not perform such a disassociation since that animation may remain "relevant" [0]. However, since
the cleanup described above occurs for animations associated with a timeline, such timeline-less
animations will not be processed and thus it's possible that an effect is destroyed but a (weak)
reference remains in the effect stack.

To address this issue, we add logic in the `KeyframeEffect` destructor to remove itself from any associated
effect stack, and add an assertion that it is indeed no longer in that stack. Additionally, we now check
in the `ElementAnimationRareData` destructor that it no longer has an effect stack or that the stack is empty.

[0] https://drafts.csswg.org/web-animations-1/#relevant-animations-section

* LayoutTests/webanimations/null-timeline-crash-expected.txt: Added.
* LayoutTests/webanimations/null-timeline-crash.html: Added.
* Source/WebCore/animation/ElementAnimationRareData.cpp:
(WebCore::ElementAnimationRareData::~ElementAnimationRareData):
* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::~KeyframeEffect):
* Source/WebCore/animation/KeyframeEffect.h:

Canonical link: https://commits.webkit.org/297297.371@safari-7622-branch