Add extra validation when deserializing NSArrays https://bugs.webkit.org/show_bug.cgi?id=298370 rdar://159206217
Reviewed by Abrar Rahman Protyasha.

The SecureCodingAllowed attribute used to tell NSKeyedUnarchiver to only allow the two types.
On new OSes, we no longer use NSKeyedUnarchiver, but we use a variant of types in CoreIPCNSCFObject.
In a few cases, this caused an increase in the types we can decode.  Adding a validator brings
that back down.  This is the cleanest way to support it while still supporting OSes where
we still use NSKeyedUnarchiver.  At some point, we'll be able to unify and tighten this even more.

* Source/WebKit/Shared/ApplePay/ApplePayPaymentSetupFeatures.serialization.in:
* Source/WebKit/Shared/Cocoa/ArgumentCodersCocoa.h:
(IPC::arrayContainsAnythingBut):
* Source/WebKit/Shared/Cocoa/DataDetectionResult.serialization.in:
* Source/WebKit/Shared/ios/InteractionInformationAtPosition.serialization.in:

Canonical link: https://commits.webkit.org/297297.374@safari-7622-branch