Prevent speechSynthesis from being accessible in LDM. https://bugs.webkit.org/show_bug.cgi?id=298382 rdar://159462101
Reviewed by Alex Christensen.

We were missing an EnabledBySetting=SpeechSynthesisAPIEnabled guard on the
DOMWindow+SpeechSynthesis.idl file. This results in window.speechSynthesis
still remaining accessible to sites, even when the setting is disabled. This
then results in process termination when the IPC message is attempted and the
privileged process marks the IPC message as invalid.

This adds the guard, and a new test to ensure this is sufficiently disabled
when in Lockdown Mode.

* LayoutTests/fast/speechsynthesis/lockdown-mode/speech-synthesis-expected.txt: Added.
* LayoutTests/fast/speechsynthesis/lockdown-mode/speech-synthesis.html: Added.
* LayoutTests/platform/glib/TestExpectations:
* LayoutTests/platform/mac-wk1/TestExpectations:
* LayoutTests/platform/win/TestExpectations:
* LayoutTests/platform/wpe/TestExpectations:
* Source/WebCore/Modules/speech/DOMWindow+SpeechSynthesis.idl:
* Source/WebCore/bindings/js/WebCoreBuiltinNames.h:

Canonical link: https://commits.webkit.org/297297.375@safari-7622-branch