[JSC] Ensure Sink+Escape+Materialize ArrayButterfly allocations are not Escaped kind rdar://159801389 https://bugs.webkit.org/show_bug.cgi?id=298356
Reviewed by Keith Miller.

When we Sink+Escape+Materialize an array, we also need to do
the same for its corresponding GetButterfly (ArrayButterfly allocation).
This is because GetButterfly cannot refer to a phantom array node - it
requires a materialized array. Therefore, when we Sink+Escape+Materialize
an ArrayButterfly allocation, we must ensure its allocation kind is
ArrayButterfly, not Escaped, so that createMaterialization() can properly
handle it.

Canonical link: https://commits.webkit.org/297297.382@safari-7622-branch