[WebKit][Main+SU] [bcdfa3892054a35c] ASAN_TRAP | WebCore::WebAnimationTime::approximatelyLessThan; WebCore::AnimationEffectTiming::getBasicTiming; WebCore::AnimationEffect::getBasicTiming https://bugs.webkit.org/show_bug.cgi?id=298869 rdar://160116687
Reviewed by Anne van Kesteren.

As part of 294049@main we started storing a "pending start time" for each animation which we would
set to be the start time when marking the animation as ready. When switching from a monotonic to a
progress-based timeline, we may be in a situation where that "pending start time" was recorded based
on the monotonic timeline, and if the animation is paused with a monotonic "pending start time" then
we will try to run some logic based on both monotonic and progress-based times in `WebAnimation::runPendingPauseTask()`.

To address this, we reset the "pending start time" when switching timelines, which makes sense regardless
of the type of timeline we're dealing with.

* LayoutTests/webanimations/pause-after-switching-to-finite-timeline-expected.txt: Added.
* LayoutTests/webanimations/pause-after-switching-to-finite-timeline.html: Added.
* Source/WebCore/animation/WebAnimation.cpp:
(WebCore::WebAnimation::setTimelineInternal):

Canonical link: https://commits.webkit.org/297297.417@safari-7622-branch