GuardMalloc crash in MediaStreamAudioSource::consumeAudio(...) https://bugs.webkit.org/show_bug.cgi?id=298971 rdar://150526896
Reviewed by Eric Carlson.

MediaStreamAudioSource::consumeAudio() will re-create its WebAudioBufferList when it detects
that the buffer list's channelCount() does not match the incoming AudioBus's numberOfChannels().
However, this is a logic error; WebAudioBufferList::channelCount() returns the number of
interleaved channels within each of the list's buffers, not the number of buffers within the list.

Fix this logic error, and also move the creation/re-creation of the WebAudioBufferList outside
of consumeAudio(), and into a new method that is called when the nodes are reconfigured, and a
lock is taken on the render thread. Also, add a number of ASSERT checks that validate assumptions
we make about the memory layout of the WebAudioBufferList we are given.

* Source/WebCore/Modules/webaudio/AudioBasicInspectorNode.h:
* Source/WebCore/Modules/webaudio/MediaStreamAudioDestinationNode.cpp:
(WebCore::MediaStreamAudioDestinationNode::checkNumberOfChannelsForInput):
* Source/WebCore/Modules/webaudio/MediaStreamAudioDestinationNode.h:
* Source/WebCore/Modules/webaudio/MediaStreamAudioSource.cpp:
(WebCore::MediaStreamAudioSource::setNumberOfChannels):
* Source/WebCore/Modules/webaudio/MediaStreamAudioSource.h:
* Source/WebCore/Modules/webaudio/MediaStreamAudioSourceCocoa.cpp:
(WebCore::copyChannelData):
(WebCore::MediaStreamAudioSource::setNumberOfChannels):
(WebCore::MediaStreamAudioSource::consumeAudio):

Canonical link: https://commits.webkit.org/297297.425@safari-7622-branch