Sandboxed iframe without same-origin flag should not have access to its site's cookies https://bugs.webkit.org/show_bug.cgi?id=286769 rdar://143051787
Reviewed by Alex Christensen.

If an iframe's sandbox does not have the SameOrigin flag, then its document
should not have access to its cookies. Currently this is not enforced when
loading the document. This patch disables cookies when fetching a document for
a sandboxed iframe.

New layout tests added for loading a sandboxed document and added a test for
same-site websocket connection.

* LayoutTests/http/tests/cookies/block-cookies-when-loading-document-in-sandboxed-iframe.https-expected.txt: Added.
* LayoutTests/http/tests/cookies/block-cookies-when-loading-document-in-sandboxed-iframe.https.html: Added.
* LayoutTests/http/tests/cookies/resources/cookie-utilities.js:
* LayoutTests/http/tests/websocket/tests/hybi/resources/set-cookie-with-websocket-and-echo.https.html: Added.
* LayoutTests/http/tests/websocket/tests/hybi/websocket-in-sandboxed-iframe.https-expected.txt: Added.
* LayoutTests/http/tests/websocket/tests/hybi/websocket-in-sandboxed-iframe.https.html: Added.
* LayoutTests/platform/mac-wk1/TestExpectations:
* Source/WebCore/loader/DocumentLoader.cpp:
(WebCore::DocumentLoader::loadMainResource):

Canonical link: https://commits.webkit.org/297297.499@safari-7622-branch