[JSC] DataViewGetByteLength resizable-array path is not setting OSR exit correctly https://bugs.webkit.org/show_bug.cgi?id=301257 rdar://160383983
Reviewed by Tadeu Zagallo.

There are two issues in the implementation of resizable-array-accepting
DataViewGetByteLength. This is because this node is specially doing OSR
exit from a patchpoint. And other patchpoints are typically (1) not
having a result value and not having a complicated computation or (2)
OSR exits only when exception is thrown and exception-related-OSR-exit
is handled in a different way. This patch fixes DataViewGetByteLength
patchpoint by introducing two changes.

1. resultGPR can be modified before jumping to OSR exit. This is
   problematic since OSR exit's stackmap is capturing what we see before
   this patchpoint starts. We should def the result only after ensuring
   that we will not do OSR exit. This patch adds one scratch register
   and storing a result to that. And after OSR exit check is done, we
   move it to the resultGPR.
2. osrExitArgumentOffset is not correct as it requires result value
   offset. Adding +1 to that offset.

Test: JSTests/stress/data-view-byte-length-oob-exit.js

* JSTests/stress/data-view-byte-length-oob-exit.js: Added.
(i.counter.else.catch):
* Source/JavaScriptCore/ftl/FTLLowerDFGToB3.cpp:
(JSC::FTL::DFG::LowerDFGToB3::typedArrayLength):

Identifier: 301765.85@safari-7623-branch