[Floats] Crash in WebCore::PODIntervalTree::searchForOverlapsFrom; WebCore::PODIntervalTree::searchForOverlapsFrom; WebCore::FloatingObjects::logicalRightOffset https://bugs.webkit.org/show_bug.cgi?id=301301 <rdar://
Reviewed by Antti Koivisto.

PODIntervalTree node needs to be "updated" when associated float position changes.

(PODIntervalTree is a red-black tree where nodes are floating objects with y-maxY intervals.
In order to be able for find such floats in the tree, we have to ensure their positions are always up-to-date.)

Test: fast/block/float/incorrect-float-state-after-align-shift.html

* LayoutTests/fast/block/float/incorrect-float-state-after-align-shift-expected.txt: Added.
* LayoutTests/fast/block/float/incorrect-float-state-after-align-shift.html: Added.
* Source/WebCore/rendering/FloatingObjects.cpp:
(WebCore::FloatingObjects::shiftFloatsBy):

Identifier: 301765.105@safari-7623-branch