com.apple.WebKit.WebContent at com.apple.WebCore: WebCore::ResourceUsageThread::platformCollectCPUData https://bugs.webkit.org/show_bug.cgi?id=301716 rdar://140592689
Reviewed by Abrar Rahman Protyasha.

In https://commits.webkit.org/209170@main, we improved the
categorization of the threads shown in the CPU profiler. To identify
a thread as used by WebKit, we rely on a set of heuristics, one of them
being the thread's current dispatch queue's name. This required first
dereferencing a dispatch_queue_t*, which is unsafe given the queue may
be destroyed concurrently.

Given there appears to be no safe alternative way of getting the queue's
name, remove the retrieval and therefore the check using it. This means
a good portion of WebKit-related threads will still be discovered and
categorized, except for those temporarily-borrowed with the Grand
Central Dispatch, which will start to be labeled as unknown/other
threads. Given the severity and frequency of the crash due to the unsafe
dereferencing, I believe it's worth it to prevent it at the expense of
the reduced functionality, and find a way to safely implement it later
(https://webkit.org/b/301777).

No new tests. The special behavior that this patch removed was not
not covered by tests, so no tests should be broken. The basic thread
identification is covered by existing tests inspector/cpu-profiler/{threads,tracking}.html,
which should continue to pass.

* Source/WebCore/page/cocoa/ResourceUsageThreadCocoa.mm:
(WebCore::threadInfos):
(WebCore::ResourceUsageThread::platformCollectCPUData):

Identifier: 301765.247@safari-7623-branch